requirements.txt
python-telegram-bot>=21,<22
SQLAlchemy>=2.0
pydantic>=2.6
pydantic-settings>=2.2
